requirements.txt
astroid==2.13.4
attrs==22.2.0
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.1.0
Faker==16.6.1
gitdb==4.0.10
GitPython==3.1.30
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pylint==2.15.10
pytest==7.2.1
pytest-pylint==0.19.0
python-dateutil==2.8.2
six==1.16.0
smmap==5.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
